Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/spec-examples/shape-outside-018-expected.txt

CONSOLE MESSAGE: line 2659: Error: assert_approx_equals: Line 0 is positioned properly expected 48 +/- 2 but got 180
The test passes if the longest green horizontal bar is beneath the triangle and the rest of them are to its right and none intersect it. There should be no red.

XXXXXXXXXX XXXXXXXX XXXXXX XXXX XXX XXXXXXXXXXXX

FAIL CSS Test: shape-margin offset from a polygonal shape-outside Error: assert_approx_equals: Line 0 is positioned properly expected 48 +/- 2 but got 180
